requirements.txt
gradio
chromadb
langchain-community
sentence-transformers
langchain-groq
langchain-google-genai
langchain-core
langchain-ollama
langchain_text_splitters
pypdf
python-dotenv
streamlit
elevenlabs
fastembed
ollama
pdfplumber
sentence-transformers
unstructured
unstructured[all-docs]
